subscriber/requirements.txt
opencv-python
paho-mqtt
flask
numpy
mediapipe